+++ lib/class-wordpress-starter-plugin-settings.php
@@ -24,7 +24,7 @@
 
 
 class WordPress_Starter_Settings {
-	const ID = 'wordpress-starter-plugin-settings';
+	const ID = 'wordpress-starter-settings';
 
 	public static $default  = array(
 		'backwards' => array(
@@ -52,7 +52,7 @@ class WordPress_Starter_Settings {
 		add_action( 'admin_init', array( &$this, 'admin_init' ) );
 		add_action( 'admin_menu', array( &$this, 'admin_menu' ) );
 		add_action( 'init', array( &$this, 'init' ) );
-		load_plugin_textdomain( 'wordpress-starter-plugin', false, '/wordpress-starter-plugin/languages/' );
+		load_plugin_textdomain( 'wordpress-starter', false, '/wordpress-starter/languages/' );
 	}
 
 
@@ -63,10 +63,10 @@ class WordPress_Starter_Settings {
 
 
 	public static function sections() {
-		self::$sections['general']   = esc_html__( 'General', 'wordpress-starter-plugin' );
-		self::$sections['post_type'] = esc_html__( 'Post Type', 'wordpress-starter-plugin' );
-		self::$sections['reset']     = esc_html__( 'Compatibility & Reset', 'wordpress-starter-plugin' );
-		self::$sections['about']     = esc_html__( 'About WordPress Starter', 'wordpress-starter-plugin' );
+		self::$sections['general']   = esc_html__( 'General', 'wordpress-starter' );
+		self::$sections['post_type'] = esc_html__( 'Post Type', 'wordpress-starter' );
+		self::$sections['reset']     = esc_html__( 'Compatibility & Reset', 'wordpress-starter' );
+		self::$sections['about']     = esc_html__( 'About WordPress Starter', 'wordpress-starter' );
 
 		self::$sections = apply_filters( 'wordpress_starter_plugin_sections', self::$sections );
 	}
@@ -80,38 +80,38 @@ class WordPress_Starter_Settings {
 	public static function settings() {
 		// General
 		self::$settings['disable_quotes'] = array(
-			'title' => esc_html__( 'Hide built-in quotes?', 'wordpress-starter-plugin' ),
-			'desc' => esc_html__( 'Remove open and close quote span tags surrounding asdf content', 'wordpress-starter-plugin' ),
+			'title' => esc_html__( 'Hide built-in quotes?', 'wordpress-starter' ),
+			'desc' => esc_html__( 'Remove open and close quote span tags surrounding asdf content', 'wordpress-starter' ),
 			'type' => 'checkbox',
 		);
 
 		self::$settings['hide_not_found'] = array(
-			'title' => esc_html__( 'Hide "Testimonials Not Found"?', 'wordpress-starter-plugin' ),
+			'title' => esc_html__( 'Hide "Testimonials Not Found"?', 'wordpress-starter' ),
 			'type' => 'checkbox',
 		);
 
 		self::$settings['paging'] = array(
-			'title' => esc_html__( 'Enable Paging?', 'wordpress-starter-plugin' ),
-			'desc' => esc_html__( 'For `[asdfswidget_list]`', 'wordpress-starter-plugin' ),
+			'title' => esc_html__( 'Enable Paging?', 'wordpress-starter' ),
+			'desc' => esc_html__( 'For `[asdfswidget_list]`', 'wordpress-starter' ),
 			'type' => 'select',
 			'choices' => array(
-				'' => esc_html__( 'Disable', 'wordpress-starter-plugin' ),
-				1 => esc_html__( 'Enable', 'wordpress-starter-plugin' ),
-				'before' => esc_html__( 'Before asdfs', 'wordpress-starter-plugin' ),
-				'after' => esc_html__( 'After asdfs', 'wordpress-starter-plugin' ),
+				'' => esc_html__( 'Disable', 'wordpress-starter' ),
+				1 => esc_html__( 'Enable', 'wordpress-starter' ),
+				'before' => esc_html__( 'Before asdfs', 'wordpress-starter' ),
+				'after' => esc_html__( 'After asdfs', 'wordpress-starter' ),
 			),
 			'std' => 1,
 			'widget' => 0,
 		);
 
 		// Post Type
-		$desc        = __( 'URL slug-name for <a href="%1s">asdfs archive</a> page.', 'wordpress-starter-plugin' );
+		$desc        = __( 'URL slug-name for <a href="%1s">asdfs archive</a> page.', 'wordpress-starter' );
 		$has_archive = cbqe_get_option( 'has_archive', '' );
 		$site_url    = site_url( '/' . $has_archive );
 
 		self::$settings['has_archive'] = array(
 			'section' => 'post_type',
-			'title' => esc_html__( 'Archive Page URL', 'wordpress-starter-plugin' ),
+			'title' => esc_html__( 'Archive Page URL', 'wordpress-starter' ),
 			'desc' => sprintf( $desc, $site_url ),
 			'std' => 'asdfs-archive',
 			'validate' => 'sanitize_title',
@@ -121,9 +121,9 @@ class WordPress_Starter_Settings {
 		// Reset
 		self::$settings['use_cpt_taxonomy'] = array(
 			'section' => 'reset',
-			'title' => esc_html__( 'Don\'t Use Default Taxonomies?', 'wordpress-starter-plugin' ),
+			'title' => esc_html__( 'Don\'t Use Default Taxonomies?', 'wordpress-starter' ),
 			'type' => 'checkbox',
-			'desc' => esc_html__( 'If checked, use WordPress Starter\'s own category and tag taxonomies instead', 'wordpress-starter-plugin' ),
+			'desc' => esc_html__( 'If checked, use WordPress Starter\'s own category and tag taxonomies instead', 'wordpress-starter' ),
 			'widget' => 0,
 		);
 
@@ -134,9 +134,9 @@ class WordPress_Starter_Settings {
 
 			self::$settings['export'] = array(
 				'section' => 'reset',
-				'title' => esc_html__( 'Export Settings', 'wordpress-starter-plugin' ),
+				'title' => esc_html__( 'Export Settings', 'wordpress-starter' ),
 				'type' => 'readonly',
-				'desc' => esc_html__( 'These are your current settings in a serialized format. Copy the contents to make a backup of your settings.', 'wordpress-starter-plugin' ),
+				'desc' => esc_html__( 'These are your current settings in a serialized format. Copy the contents to make a backup of your settings.', 'wordpress-starter' ),
 				'std' => $serialized_options,
 				'widget' => 0,
 			);
@@ -144,27 +144,27 @@ class WordPress_Starter_Settings {
 
 		self::$settings['import'] = array(
 			'section' => 'reset',
-			'title' => esc_html__( 'Import Settings', 'wordpress-starter-plugin' ),
+			'title' => esc_html__( 'Import Settings', 'wordpress-starter' ),
 			'type' => 'textarea',
-			'desc' => esc_html__( 'Paste new serialized settings here to overwrite your current configuration.', 'wordpress-starter-plugin' ),
+			'desc' => esc_html__( 'Paste new serialized settings here to overwrite your current configuration.', 'wordpress-starter' ),
 			'widget' => 0,
 		);
 
 		self::$settings['delete_data'] = array(
 			'section' => 'reset',
-			'title' => esc_html__( 'Remove Plugin Data on Deletion?', 'wordpress-starter-plugin' ),
+			'title' => esc_html__( 'Remove Plugin Data on Deletion?', 'wordpress-starter' ),
 			'type' => 'checkbox',
 			'class' => 'warning', // Custom class for CSS
-			'desc' => esc_html__( 'Delete all WordPress Starter data and options from database on plugin deletion', 'wordpress-starter-plugin' ),
+			'desc' => esc_html__( 'Delete all WordPress Starter data and options from database on plugin deletion', 'wordpress-starter' ),
 			'widget' => 0,
 		);
 
 		self::$settings['reset_defaults'] = array(
 			'section' => 'reset',
-			'title' => esc_html__( 'Reset to Defaults?', 'wordpress-starter-plugin' ),
+			'title' => esc_html__( 'Reset to Defaults?', 'wordpress-starter' ),
 			'type' => 'checkbox',
 			'class' => 'warning', // Custom class for CSS
-			'desc' => esc_html__( 'Check this box to reset options to their defaults', 'wordpress-starter-plugin' ),
+			'desc' => esc_html__( 'Check this box to reset options to their defaults', 'wordpress-starter' ),
 			'widget' => 0,
 		);
 
@@ -225,7 +225,7 @@ class WordPress_Starter_Settings {
 
 
 	public function admin_menu() {
-		$admin_page = add_options_page( esc_html__( 'WordPress Starter Settings', 'wordpress-starter-plugin' ), esc_html__( 'Custom Bulk/Quick', 'wordpress-starter-plugin' ), 'manage_options', self::ID, array( 'WordPress_Starter_Settings', 'display_page' ) );
+		$admin_page = add_options_page( esc_html__( 'WordPress Starter Settings', 'wordpress-starter' ), esc_html__( 'Custom Bulk/Quick', 'wordpress-starter' ), 'manage_options', self::ID, array( 'WordPress_Starter_Settings', 'display_page' ) );
 
 		add_action( 'admin_print_scripts-' . $admin_page, array( &$this, 'scripts' ) );
 		add_action( 'admin_print_styles-' . $admin_page, array( &$this, 'styles' ) );
@@ -262,7 +262,7 @@ class WordPress_Starter_Settings {
 	public static function display_page() {
 		echo '<div class="wrap">
 			<div class="icon32" id="icon-options-general"></div>
-			<h2>' . esc_html__( 'WordPress Starter Settings', 'wordpress-starter-plugin' ) . '</h2>';
+			<h2>' . esc_html__( 'WordPress Starter Settings', 'wordpress-starter' ) . '</h2>';
 
 		echo '<form action="options.php" method="post">';
 
@@ -279,16 +279,16 @@ class WordPress_Starter_Settings {
 		self::do_settings_sections( self::ID );
 
 		echo '
-			<p class="submit"><input name="Submit" type="submit" class="button-primary" value="' . esc_html__( 'Save Changes', 'wordpress-starter-plugin' ) . '" /></p>
+			<p class="submit"><input name="Submit" type="submit" class="button-primary" value="' . esc_html__( 'Save Changes', 'wordpress-starter' ) . '" /></p>
 			</form>
 		</div>
 		';
 
 		echo '
-			<p>If you like this plugin, please <a href="http://aihr.us/about-aihrus/donate/" title="Donate for Good Karma"><img src="https://www.paypalobjects.com/en_US/i/btn/btn_donate_SM.gif" border="0" alt="Donate for Good Karma" /></a> or <a href="http://aihr.us/downloads/wordpress-starter-plugin-premium-wordpress-plugin/" title="purchase WordPress Starter Premium">purchase WordPress Starter Premium</a> to help fund further development and <a href="http://wordpress.org/support/plugin/wordpress-starter-plugin" title="Support forums">support</a>.</p>
+			<p>If you like this plugin, please <a href="http://aihr.us/about-aihrus/donate/" title="Donate for Good Karma"><img src="https://www.paypalobjects.com/en_US/i/btn/btn_donate_SM.gif" border="0" alt="Donate for Good Karma" /></a> or <a href="http://aihr.us/downloads/wordpress-starter-premium-wordpress-plugin/" title="purchase WordPress Starter Premium">purchase WordPress Starter Premium</a> to help fund further development and <a href="http://wordpress.org/support/plugin/wordpress-starter" title="Support forums">support</a>.</p>
 		';
 
-		$text = esc_html__( 'Copyright &copy;%1$s %2$s.', 'wordpress-starter-plugin' );
+		$text = esc_html__( 'Copyright &copy;%1$s %2$s.', 'wordpress-starter' );
 		$link = '<a href="http://aihr.us">Aihrus</a>';
 		echo '<p class="copyright">' . sprintf( $text, date( 'Y' ), $link ) . '</p>';
 
@@ -342,7 +342,7 @@ class WordPress_Starter_Settings {
 	public function display_about_section() {
 		echo '
 			<div id="about" style="width: 70%; min-height: 225px;">
-				<p><img class="alignright size-medium" title="Michael in Red Square, Moscow, Russia" src="/wp-content/plugins/wordpress-starter-plugin/media/michael-cannon-red-square-300x2251.jpg" alt="Michael in Red Square, Moscow, Russia" width="300" height="225" /><a href="http://wordpress.org/extend/plugins/wordpress-starter-plugin/">WordPress Starter</a> is by <a href="http://aihr.us/about-aihrus/michael-cannon-resume/">Michael Cannon</a>. He\'s <a title="Lot\'s of stuff about Peichi Liu…" href="http://peimic.com/t/peichi-liu/">Peichi’s</a> smiling man, an adventurous <a title="Water rat" href="http://www.chinesehoroscope.org/chinese_zodiac/rat/" target="_blank">water-rat</a>, <a title="Axelerant – Open Source. Engineered." href="http://axelerant.com/who-we-are">chief people officer</a>, <a title="Aihrus – website support made easy since 1999" href="http://aihr.us/about-aihrus/">chief technology officer</a>, <a title="Road biker, cyclist, biking; whatever you call, I love to ride" href="http://peimic.com/c/biking/">cyclist</a>, <a title="Michael\'s poetic like literary ramblings" href="http://peimic.com/t/poetry/">poet</a>, <a title="World Wide Opportunities on Organic Farms" href="http://peimic.com/t/WWOOF/">WWOOF’er</a> and <a title="My traveled to country list, is more than my age." href="http://peimic.com/c/travel/">world traveler</a>.</p>
+				<p><img class="alignright size-medium" title="Michael in Red Square, Moscow, Russia" src="/wp-content/plugins/wordpress-starter/media/michael-cannon-red-square-300x2251.jpg" alt="Michael in Red Square, Moscow, Russia" width="300" height="225" /><a href="http://wordpress.org/extend/plugins/wordpress-starter/">WordPress Starter</a> is by <a href="http://aihr.us/about-aihrus/michael-cannon-resume/">Michael Cannon</a>. He\'s <a title="Lot\'s of stuff about Peichi Liu…" href="http://peimic.com/t/peichi-liu/">Peichi’s</a> smiling man, an adventurous <a title="Water rat" href="http://www.chinesehoroscope.org/chinese_zodiac/rat/" target="_blank">water-rat</a>, <a title="Axelerant – Open Source. Engineered." href="http://axelerant.com/who-we-are">chief people officer</a>, <a title="Aihrus – website support made easy since 1999" href="http://aihr.us/about-aihrus/">chief technology officer</a>, <a title="Road biker, cyclist, biking; whatever you call, I love to ride" href="http://peimic.com/c/biking/">cyclist</a>, <a title="Michael\'s poetic like literary ramblings" href="http://peimic.com/t/poetry/">poet</a>, <a title="World Wide Opportunities on Organic Farms" href="http://peimic.com/t/WWOOF/">WWOOF’er</a> and <a title="My traveled to country list, is more than my age." href="http://peimic.com/c/travel/">world traveler</a>.</p>
 			</div>
 		';
 	}
@@ -648,7 +648,7 @@ class WordPress_Starter_Settings {
 
 		case 'required':
 			if ( empty( $input[ $id ] ) )
-				$errors[ $id ] = esc_html__( 'Required', 'wordpress-starter-plugin' );
+				$errors[ $id ] = esc_html__( 'Required', 'wordpress-starter' );
 			break;
 
 		case 'slug':