+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG WordPress Starter plugin for WordPress
 
 ## TBD
+* Baseline API
+* Delete prepare script on run
+* Purge excess code
+* Update POT
 
 ## 0.0.1
 * Initial code release